This project is a web application that uses WebGL (Web Graphics Library) to create the Turkish flag. The project determines the background color of the Turkish flag and draws a white star and a red crescent shape on it. WebGL technology enables fast and interactive 2D and 3D graphics rendering within web browsers.
This project aims to create the symbols and colors of the Turkish flag using WebGL technology. The project includes the distinctive elements of the Turkish flag, which are a red crescent, a white star, and the background color. Additionally, separate code snippets are provided to generate each symbol.

The purpose of this project is to showcase the power of WebGL technology and demonstrate how impressive graphics can be achieved within web browsers. WebGL support in the browser is required to run the application. By adjusting the parameters for the white star, red crescent, and background color, the application creates the Turkish flag.

Users can download and run this project or experience it directly in their web browsers. By doing so, they can visually experience the symbols and colors of the Turkish flag. This project can serve as a tutorial resource for developers looking to understand the potential of WebGL technology and delve into browser-based graphics programming.